feat(ui/downloads): create DownloadsContent component for managing downloads with sorting, pagination, and URL parameter handling - This was done to work around a hydration error caused by useSearchParams()
